break apart logic for various jobs in worker module
