+++ packages/host/tests/helpers/adapter.ts
@@ -13,9 +13,9 @@ import {
   hasExecutableExtension,
   Deferred,
   unixTime,
+  type LintResult,
 } from '@cardstack/runtime-common';
 
-import type { LintResult } from '@cardstack/runtime-common/lint';
 import type { MatrixClient } from '@cardstack/runtime-common/matrix-client';
 import { APP_BOXEL_REALM_EVENT_TYPE } from '@cardstack/runtime-common/matrix-constants';
